requirements.txt
absl-py==2.3.0
aiosqlite==0.21.0
alembic==1.16.1
annotated-types==0.7.0
anyio==3.7.1
apprise==1.9.3
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgi-lifespan==2.1.0
astroid==3.3.10
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
coolname==2.2.0
croniter==2.0.7
cryptography==45.0.3
cycler==0.12.1
databricks-cli==0.18.0
dateparser==1.2.0
db-dtypes==1.4.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
dnspython==2.7.0
docker==6.1.3
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
Flask==2.3.3
flatbuffers==25.2.10
fonttools==4.58.1
fqdn==1.5.1
fsspec==2025.5.1
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.8.2
google-auth==2.40.2
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.18.0
google-cloud-core==2.4.3
google-cloud-storage==2.14.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.56.4
graphviz==0.20.3
greenlet==3.2.2
griffe==1.7.3
grpcio==1.72.1
grpcio-status==1.62.3
gunicorn==20.1.0
h11==0.16.0
h2==4.2.0
h5py==3.10.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
importlib-metadata==5.2.0
importlib_resources==6.1.3
iniconfig==2.1.0
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.37.0
isoduration==20.11.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
joblib==1.5.1
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
kubernetes==29.0.0
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
mlflow==2.1.1
namex==0.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nbresult==0.1.1
nest-asyncio==1.6.0
networkx==3.4.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.23.5
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.16.0
orjson==3.10.18
overrides==7.7.0
packaging==22.0
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pendulum==2.1.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
prefect==2.19.2
prometheus_client==0.22.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==3.19.6
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==10.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pygeohash==3.1.3
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.7
pyparsing==3.2.3
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2022.7.1
pytzdata==2020.1
PyYAML==6.0.2
pyzmq==26.4.0
querystring-parser==1.2.4
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.18.12
ruamel.yaml.clib==0.2.12
scikit-learn==1.3.1
scipy==1.10.0
seaborn==0.13.2
Send2Trash==1.8.3
shap==0.47.2
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
-e git+ssh://git@github.com/mc-will/BYU_Locating-Bacterial_Flagellar_Motors_2025.git@ae86cb6018288332cf82064427511a497078a1a5#egg=source
SQLAlchemy==1.4.54
sqlparse==0.5.3
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tifffile==2025.5.10
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.12.5
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.28.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.2
yapf==0.43.0
zipp==3.22.0
